# web/docs/openapi/qoz-platform.openapi.yaml
openapi: 3.1.0
info:
  title: Qoz Platform API
  version: 1.0.0
  description: |
    Single spec for qoz-vision-prod-web: Section A documents live Next.js routes;
    Section B defines target contracts for mock domains (Repository layer).
servers:
  - url: http://localhost:3000
    description: Next.js (qoz-vision-prod-web)
  - url: http://localhost:8080
    description: Backend (qoz-vision-prod-backend)
...
paths:
...
  /api/incidents/{id}:
    get:
      tags: [Incidents]
      summary: Get incident (live)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Incident detail
...